Resolve dependencies of and to Decoration
Move the Theme to an extra file theme.h/.cpp annd make the
Decoration-member of Client a pointer.
