+++ src/clientmanager.cpp
@@ -5,6 +5,7 @@
 
 #include "client.h"
 #include "completion.h"
+#include "decoration.h"
 #include "ewmh.h"
 #include "frametree.h"
 #include "globals.h"
@@ -151,7 +152,7 @@ Client* ClientManager::manage_client(Window win, bool visible_already) {
     }
 
     // actually manage it
-    client->dec.createWindow();
+    client->dec->createWindow();
     client->fuzzy_fix_initial_position();
     add(client);
     // insert to layout
@@ -229,7 +230,7 @@ void ClientManager::force_unmanage(Client* client) {
     XSelectInput(g_display, client->window_, 0);
     //XUngrabButton(g_display, AnyButton, AnyModifier, win);
     // permanently remove it
-    XUnmapWindow(g_display, client->dec.decorationWindow());
+    XUnmapWindow(g_display, client->decorationWindow());
     XReparentWindow(g_display, client->window_, g_root, 0, 0);
     client->clear_properties();
     HSTag* tag = client->tag();
